requirements.txt
streamlit
pandas
openpyxl
plotly
google-genai>=1.30.0
